third_party/blink/web_tests/http/tests/images/document-policy-unoptimized-images-cached-image-expected.txt

CONSOLE ERROR: Document policy violation: lossy-images-max-bpp is not allowed in this document.
CONSOLE ERROR: Document policy violation: lossy-images-max-bpp is not allowed in this document.
CONSOLE ERROR: Document policy violation: lossy-images-max-bpp is not allowed in this document.
